requirements.txt
Flask>=3.0.0
transformers>=4.44.0
accelerate>=0.33.0
langdetect>=1.0.9
gunicorn>=21.2.0
huggingface-hub>=0.23.0
numpy>=1.26.0
python-dotenv>=1.0.1
gTTS>=2.5.0
